Fix onboarding Add type to contract
